requirements.txt
Flask
elasticsearch>=5.0.0,<6.0.0
pytz
python-dateutil
requests
